requirements.txt
absl-py==0.7.1
alembic==1.0.10
aniso8601==6.0.0
APScheduler==3.6.0
apturl==0.5.2
arabic-reshaper==2.0.14
asn1crypto==0.24.0
astor==0.7.1
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
awscli==1.16.144
beautifulsoup4==4.6.0
bleach==3.1.0
botocore==1.12.134
Brlapi==0.6.6
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.3.9
command-not-found==0.3
configparser==3.7.4
coverage==4.5.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==2.2.2
docutils==0.14
entrypoints==0.3
flake8==3.7.7
Flask==1.0.3
flask-marshmallow==0.10.1
Flask-Migrate==2.5.1
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
future==0.17.1
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
httplib2==0.9.2
idna==2.7
imageio==2.5.0
imgaug==0.2.9
importlib-metadata==0.18
iso8601==0.1.12
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.10
Markdown==3.1
MarkupSafe==1.1.1
marshmallow==2.19.2
marshmallow-sqlalchemy==0.16.3
matplotlib==3.0.3
mccabe==0.6.1
memory-profiler==0.55.0
mock==2.0.0
more-itertools==7.0.0
mysqlclient==1.4.2.post1
netifaces==0.10.4
networkx==2.3
numpy==1.16.3
oauth==1.0.1
olefile==0.45.1
opencv-python==4.1.0.25
packaging==19.0
paho-mqtt==1.4.0
pbr==5.1.3
pdf2image==1.5.1
pexpect==4.2.1
Pillow==6.0.0
pkginfo==1.5.0.1
pluggy==0.12.0
protobuf==3.7.1
psutil==5.6.2
py==1.8.0
pycairo==1.16.2
pycodestyle==2.5.0
pycrypto==2.6.1
pycups==1.9.73
pyflakes==2.1.1
Pygments==2.3.1
pygobject==3.26.1
pylint==2.3.1
pylint-flask==0.6
pylint-plugin-utils==0.5
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.0
pyRFC3339==1.0
pyserial==3.4
pytest==4.6.3
pytest-cov==2.7.1
python-apt==1.6.3+ubuntu1
python-bidi==0.4.0
python-dateutil==2.8.0
python-debian==0.1.32
python-editor==1.0.4
pytz==2019.1
PyWavelets==1.0.3
pyxdg==0.25
PyYAML==5.1.1
readme-renderer==24.0
reportlab==3.4.0
requests==2.20.0
requests-toolbelt==0.9.1
requests-unixsocket==0.1.5
s3transfer==0.2.0
scikit-image==0.15.0
scipy==1.2.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Shapely==1.6.4.post2
simplejson==3.13.2
six==1.12.0
SQLAlchemy==1.3.3
sqlparse==0.3.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tqdm==4.23.4
twine==1.13.0
typed-ast==1.3.5
tzlocal==1.5.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.2
usb-creator==0.3.3
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
wrapt==1.11.1
xkit==0.0.0
zipp==0.5.1
zope.interface==4.3.2
